Add the option to set an inner background color
This is useful when fetching an image from the Web. Imagine the
following scenario:

1. Setup the AMPAvatarView without an image
2. Add a spinner in the middle of the avatar view
3. Set the image once it is fetched

Without the inner background color, on step 2, the border would not
be visible if it's white, and when the image is finally fetched, there
is a border all of a sudden. The inner background color solves this
problem.